+++ modules/module-mysql/src/api/MySQLRouteAPIAdapter.ts
@@ -1,4 +1,4 @@
-import { api, ParseSyncRulesOptions, storage } from '@powersync/service-core';
+import { api, ParseSyncRulesOptions, ReplicationHeadCallback, storage } from '@powersync/service-core';
 
 import * as sync_rules from '@powersync/service-sync-rules';
 import * as service_types from '@powersync/service-types';
@@ -273,6 +273,15 @@ export class MySQLRouteAPIAdapter implements api.RouteAPI {
     const result = await common.readExecutedGtid(connection);
     connection.release();
     return result.comparable;
+  }
+
+  async createReplicationHead<T>(callback: ReplicationHeadCallback<T>): Promise<T> {
+    const head = await this.getReplicationHead();
+    const r = await callback(head);
+
+    // TODO: make sure another message is replicated
+
+    return r;
   }
 
   async getConnectionSchema(): Promise<service_types.DatabaseSchema[]> {